Add helper base classes for infrared consumers
