+++ homeassistant/components/lg_infrared/button.py
@@ -5,6 +5,7 @@ from dataclasses import dataclass
 from infrared_protocols.codes.lg.tv import LGTVCode
 
 from homeassistant.components.button import ButtonEntity, ButtonEntityDescription
+from homeassistant.components.infrared import InfraredEmitterConsumerEntity
 from homeassistant.config_entries import ConfigEntry
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity_platform import AddConfigEntryEntitiesCallback
@@ -130,7 +131,7 @@ async def async_setup_entry(
         )
 
 
-class LgIrButton(LgIrEntity, ButtonEntity):
+class LgIrButton(LgIrEntity, InfraredEmitterConsumerEntity, ButtonEntity):
     """LG IR button entity."""
 
     entity_description: LgIrButtonEntityDescription
@@ -142,9 +143,10 @@ class LgIrButton(LgIrEntity, ButtonEntity):
         description: LgIrButtonEntityDescription,
     ) -> None:
         """Initialize LG IR button."""
-        super().__init__(entry, infrared_entity_id, unique_id_suffix=description.key)
+        super().__init__(entry, unique_id_suffix=description.key)
+        self._infrared_emitter_entity_id = infrared_entity_id
         self.entity_description = description
 
     async def async_press(self) -> None:
         """Press the button."""
-        await self._send_command(self.entity_description.command_code)
+        await self._send_command(self.entity_description.command_code.to_command())
