+++ homeassistant/components/lg_infrared/entity.py
@@ -1,75 +1,20 @@
 """Common entity for LG IR integration."""
 
-import logging
-
-from infrared_protocols.codes.lg.tv import LGTVCode
-
-from homeassistant.components.infrared import async_send_command
 from homeassistant.config_entries import ConfigEntry
-from homeassistant.const import STATE_UNAVAILABLE
-from homeassistant.core import Event, EventStateChangedData, callback
 from homeassistant.helpers.device_registry import DeviceInfo
 from homeassistant.helpers.entity import Entity
-from homeassistant.helpers.event import async_track_state_change_event
 
 from .const import DOMAIN
 
-_LOGGER = logging.getLogger(__name__)
-
 
 class LgIrEntity(Entity):
-    """LG IR base entity."""
+    """LG IR base entity providing common device info."""
 
     _attr_has_entity_name = True
-    _attr_should_poll = False
 
-    def __init__(
-        self, entry: ConfigEntry, infrared_entity_id: str, unique_id_suffix: str
-    ) -> None:
+    def __init__(self, entry: ConfigEntry, unique_id_suffix: str) -> None:
         """Initialize LG IR entity."""
-        self._infrared_entity_id = infrared_entity_id
         self._attr_unique_id = f"{entry.entry_id}_{unique_id_suffix}"
         self._attr_device_info = DeviceInfo(
             identifiers={(DOMAIN, entry.entry_id)}, name="LG TV", manufacturer="LG"
         )
-
-    async def async_added_to_hass(self) -> None:
-        """Subscribe to infrared entity state changes."""
-        await super().async_added_to_hass()
-
-        self.async_on_remove(
-            async_track_state_change_event(
-                self.hass, [self._infrared_entity_id], self._async_ir_state_changed
-            )
-        )
-
-        # Set initial availability based on current infrared entity state
-        ir_state = self.hass.states.get(self._infrared_entity_id)
-        self._attr_available = (
-            ir_state is not None and ir_state.state != STATE_UNAVAILABLE
-        )
-
-    async def _send_command(self, code: LGTVCode) -> None:
-        """Send an IR command using the LG protocol."""
-        await async_send_command(
-            self.hass,
-            self._infrared_entity_id,
-            code.to_command(),
-            context=self._context,
-        )
-
-    @callback
-    def _async_ir_state_changed(self, event: Event[EventStateChangedData]) -> None:
-        """Handle infrared entity state changes."""
-        new_state = event.data["new_state"]
-        ir_available = new_state is not None and new_state.state != STATE_UNAVAILABLE
-        if ir_available != self.available:
-            _LOGGER.info(
-                "Infrared entity %s used by %s is %s",
-                self._infrared_entity_id,
-                self.entity_id,
-                "available" if ir_available else "unavailable",
-            )
-
-            self._attr_available = ir_available
-            self.async_write_ha_state()
