+++ homeassistant/components/lg_infrared/media_player.py
@@ -2,6 +2,7 @@
 
 from infrared_protocols.codes.lg.tv import LGTVCode
 
+from homeassistant.components.infrared import InfraredEmitterConsumerEntity
 from homeassistant.components.media_player import (
     MediaPlayerDeviceClass,
     MediaPlayerEntity,
@@ -32,7 +33,7 @@ async def async_setup_entry(
         async_add_entities([LgIrTvMediaPlayer(entry, infrared_entity_id)])
 
 
-class LgIrTvMediaPlayer(LgIrEntity, MediaPlayerEntity):
+class LgIrTvMediaPlayer(LgIrEntity, InfraredEmitterConsumerEntity, MediaPlayerEntity):
     """LG IR media player entity."""
 
     _attr_name = None
@@ -52,45 +53,46 @@ class LgIrTvMediaPlayer(LgIrEntity, MediaPlayerEntity):
 
     def __init__(self, entry: ConfigEntry, infrared_entity_id: str) -> None:
         """Initialize LG IR media player."""
-        super().__init__(entry, infrared_entity_id, unique_id_suffix="media_player")
+        super().__init__(entry, unique_id_suffix="media_player")
+        self._infrared_emitter_entity_id = infrared_entity_id
         self._attr_state = MediaPlayerState.ON
 
     async def async_turn_on(self) -> None:
         """Turn on the TV."""
-        await self._send_command(LGTVCode.POWER_ON)
+        await self._send_command(LGTVCode.POWER_ON.to_command())
 
     async def async_turn_off(self) -> None:
         """Turn off the TV."""
-        await self._send_command(LGTVCode.POWER_OFF)
+        await self._send_command(LGTVCode.POWER_OFF.to_command())
 
     async def async_volume_up(self) -> None:
         """Send volume up command."""
-        await self._send_command(LGTVCode.VOLUME_UP)
+        await self._send_command(LGTVCode.VOLUME_UP.to_command())
 
     async def async_volume_down(self) -> None:
         """Send volume down command."""
-        await self._send_command(LGTVCode.VOLUME_DOWN)
+        await self._send_command(LGTVCode.VOLUME_DOWN.to_command())
 
     async def async_mute_volume(self, mute: bool) -> None:
         """Send mute command."""
-        await self._send_command(LGTVCode.MUTE)
+        await self._send_command(LGTVCode.MUTE.to_command())
 
     async def async_media_next_track(self) -> None:
         """Send channel up command."""
-        await self._send_command(LGTVCode.CHANNEL_UP)
+        await self._send_command(LGTVCode.CHANNEL_UP.to_command())
 
     async def async_media_previous_track(self) -> None:
         """Send channel down command."""
-        await self._send_command(LGTVCode.CHANNEL_DOWN)
+        await self._send_command(LGTVCode.CHANNEL_DOWN.to_command())
 
     async def async_media_play(self) -> None:
         """Send play command."""
-        await self._send_command(LGTVCode.PLAY)
+        await self._send_command(LGTVCode.PLAY.to_command())
 
     async def async_media_pause(self) -> None:
         """Send pause command."""
-        await self._send_command(LGTVCode.PAUSE)
+        await self._send_command(LGTVCode.PAUSE.to_command())
 
     async def async_media_stop(self) -> None:
         """Send stop command."""
-        await self._send_command(LGTVCode.STOP)
+        await self._send_command(LGTVCode.STOP.to_command())
